pyproject.toml
[project]
name = "mp-dwh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.11.4",
    "dbt-duckdb>=1.10.0",
    "pandas>=3.0.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.0",
    "pytest>=9.0.2",
    "rich>=14.3.2",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
]
